stbt-docker: Ensure that uid is the same inside and outside the container
We share the test-pack between the host and the container using a host
volume.  The UID and GID of the files in the test-pack will typically
match the UID and GID of the user on the host system.

The user within the container is "stb-tester".  Previously it had fixed
UID and GID of 1000:1000.  If the user on the host had a UID other than
1000 we would either not have permissions to read files in the test-pack,
or would create files which would be inaccessible to the user on the
host.

With this change we dynamically change the UID of the stb-tester user
to match the UID on the host to resolve these issues.