Establish stbt-docker as its own project, relicense to MIT
Prior to this commit all the commits were made in the stb-tester repo.  As
stbt-docker has a seperate release cycle to stb-tester proper and cannot
reuse any of its code it makes sense to have it separate.

This commit pulls in some of the build infrastructure from stb-tester into
this project as is necessary to test stbt-docker.  It also adds the MIT
licence, distinct to stb-tester's LGPLv2+.  This is to allow people to
include stbt-docker in their own projects git repos without concern.