Fix `stbt.ocr` when running stbt-docker with uid != 1000
`stbt.ocr` works by writing files into `$XDG_RUNTIME_DIR` allowing them
to be passed to tesseract for processing.  The test-pack entrypoint sets
this to `/run/users/1000` which has owner 1000 (stb-tester).  This is
correct if your UID is 1000, but incorrect otherwise.  This causes
`stbt.ocr` to fail with:

    OSError: [Errno 13] Permission denied: '/run/user/1000/stbt-ocr-.......'

This commit fixes this issue by creating `/run/users/{uid}` instead with
the owner matching the user that stbt-docker is running as and setting
`$XDG_RUNTIME_DIR` appropriately.