Don't create dummy /dev/raw1394 if /dev mounted from host
We create `/dev/raw1394` (a symlink pointing at `/dev/null`) to silence
an error every time a Python script imports stbt. However we don't want
to create this symlink on the host if `/dev` is mounted in from the
host!

In my experiments (with both docker 1.6.2 and 1.12.3) `/dev` is devtmpfs
if you're mounting in `/dev` from the host, otherwise it's tmpfs:

    $ docker run --rm -ti ubuntu:14.04 cat /proc/mounts | grep devtmpfs

    $ docker run --rm -ti -v /dev:/dev ubuntu:14.04 cat /proc/mounts | grep devtmpfs
    udev /dev devtmpfs rw,nosuid,relatime,size=8149048k,nr_inodes=2037262,mode=755 0 0
    udev /proc/kcore devtmpfs rw,nosuid,relatime,size=8149048k,nr_inodes=2037262,mode=755 0 0
    udev /proc/timer_list devtmpfs rw,nosuid,relatime,size=8149048k,nr_inodes=2037262,mode=755 0 0
    udev /proc/timer_stats devtmpfs rw,nosuid,relatime,size=8149048k,nr_inodes=2037262,mode=755 0 0
    udev /proc/sched_debug devtmpfs rw,nosuid,relatime,size=8149048k,nr_inodes=2037262,mode=755 0 0

    $ docker run --rm -ti -v /dev:/dev --privileged ubuntu:14.04 cat /proc/mounts | grep devtmpfs
    udev /dev devtmpfs rw,nosuid,relatime,size=8149048k,nr_inodes=2037262,mode=755 0 0

Note that `{{ ... }}` is to escape Python's string `format` operation.
The shell sees `{ ... }`.